requirements.txt
torch 
transformers 
pytorch_lightning==1.9.5 
datasets
neptune
matplotlib==3.7.1
sklearn-pandas==2.2.0
python-dotenv